requirements.txt
bcrypt<4.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
aiosqlite>=0.19.0
email-validator>=2.1.0
pytest>=8.0.0
httpx>=0.26.0
aiofiles>=23.2.1
